pyproject.toml
[project]
name = "sloths-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sloths_inventory.settings"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=src --cov-report=term-missing"
testpaths = ["src"]
